+++ tests/test_web.py
@@ -2,16 +2,12 @@
 
 """Tests for the simple web service."""
 
-import json
 import unittest
-from urllib.parse import quote
 
 from fastapi.testclient import TestClient
 
 from curies import Converter
-from curies.mapping_service import get_flask_mapping_app
 from curies.web import FAILURE_CODE, get_fastapi_app, get_flask_app
-from tests.test_mapping_service import EXPECTED, PREFIX_MAP, SPARQL_FROM_SERVICE, SPARQL_SIMPLE
 
 
 class ConverterMixin(unittest.TestCase):
@@ -69,24 +65,3 @@ class TestFlaskBlueprint(ConverterMixin):
         with self.app.test_client() as client:
             res = client.get("/NOPREFIX:NOIDENTIFIER", follow_redirects=False)
             self.assertEqual(FAILURE_CODE, res.status_code, msg=res.text)
-
-
-class TestMappingWeb(unittest.TestCase):
-    """Test the web component of the mapping service."""
-
-    def setUp(self) -> None:
-        """Set up the test case with a converter and app."""
-        self.converter = Converter.from_priority_prefix_map(PREFIX_MAP)
-        self.app = get_flask_mapping_app(self.converter)
-
-    def test_query(self):
-        """Test querying the app."""
-        for sparql in [SPARQL_SIMPLE, SPARQL_FROM_SERVICE]:
-            with self.subTest(sparql=sparql), self.app.test_client() as client:
-                res = client.get(f"/sparql?query={quote(sparql)}")
-                self.assertEqual(200, res.status_code, msg=f"Response: {res}")
-                records = {
-                    (record["s"]["value"], record["o"]["value"])
-                    for record in json.loads(res.text)["results"]["bindings"]
-                }
-                self.assertEqual(EXPECTED, records)
